validators index and show
