+++ app/controllers/api/v1/api_controller.rb
@@ -123,43 +123,6 @@ module Api
           :payload
         )
       end
-
-      private
-
-      def create_json_result(validator, with_history = false)
-        hash = {}
-
-        hash.merge!(validator.to_builder.attributes!)
-
-        score = validator.score
-        ip = score.ip_for_api if score
-        vote_account = validator.vote_accounts.last
-        validator_history = validator.most_recent_epoch_credits_by_account
-
-        
-        hash.merge!(score.to_builder(with_history: with_history).attributes!)
-        hash.merge!(ip.to_builder.attributes!) unless ip.blank?
-        hash.merge!(vote_account.to_builder.attributes!) unless vote_account.blank?
-        hash.merge!(validator_history.to_builder.attributes!) unless validator_history.blank?
-
-        # Data from the skipped_slots_report
-        unless @skipped_slots_report.nil?
-          this_report = @skipped_slots_report.payload.select do |ssr|
-            ssr['account'] == validator.account
-          end
-
-          unless this_report.first.nil?
-            hash.merge!({
-              'skipped_slots' => this_report.first['skipped_slots'],
-              'skipped_slot_percent' => this_report.first['skipped_slot_percent'],
-              'ping_time' => nil # this_report.first['ping_time']
-            })
-          end
-        end
-
-        hash.merge!({url: validator.api_url})
-        hash
-      end
     end
   end
 end
