+++ app/controllers/api/v1/validators_controller.rb
@@ -1,22 +1,18 @@
 module Api
   module V1
     class ValidatorsController < BaseController
-      def validators_list
-        @validators = ValidatorQuery.new(
-          network: "mainnet",
-          sort_order: "score",
-          limit: 9999,
-          page: 1
-        ).call
+      include ValidatorsControllerHelper
 
-        unless params[:q].blank?
-          @validators = ValidatorSearchQuery.new(@validators).search(params[:q])
-        end
+      before_action :set_skipped_slots_report
 
-        @skipped_slots_report = Report.where(
+      def index
+        @validators = ValidatorQuery.new.call(
           network: params[:network],
-          name: 'build_skipped_slot_percent'
-        ).last
+          sort_order: params[:order],
+          page: params[:page],
+          limit: params[:limit],
+          query: params[:q]
+        )
 
         json_result = @validators.map { |val| create_json_result(val) }
 
@@ -25,20 +21,15 @@ module Api
         render json: { 'status' => 'Parameter Missing' }, status: 400
       rescue StandardError => e
         Appsignal.send_error(e)
-        render json: { 'status' => e.message }, status: 500
+        render json: { 'status' => 'server error' }, status: 500
       end
 
-      def validators_show
+      def show
         @validator = Validator.select(validator_fields, validator_score_v1_fields)
                               .eager_load(:validator_score_v1)
                               .find_by(network: params[:network], account: params['account'])
 
         raise ValidatorNotFound if @validator.nil?
-
-        @skipped_slots_report = Report.where(
-          network: params[:network],
-          name: 'build_skipped_slot_percent'
-        ).last
 
         current_epoch = EpochHistory.last
 
@@ -49,30 +40,20 @@ module Api
         render json: { 'status' => 'Parameter Missing' }, status: 400
       rescue StandardError => e
         Appsignal.send_error(e)
-        render json: { 'status' => e.message }, status: 500
+        render json: { 'status' => 'server error' }, status: 500
       end
 
-      def validator_block_history
-        @validator = Validator.where(
-          network: params[:network], account: params['account']
-        ).order('network, account').first
+      private
 
-        raise ValidatorNotFound if @validator.nil?
+      def validator_params
+        params.permit(:network, :order, :page, :limit, :q)
+      end
 
-        @limit = params[:limit] || 9999
-
-        @block_history = @validator.validator_block_histories
-                                   .order('id desc')
-                                   .limit(@limit)
-
-        render 'api/v1/validators/block_history', formats: :json
-      rescue ValidatorNotFound
-        render json: { 'status' => 'Validator Not Found' }, status: 404
-      rescue ActionController::ParameterMissing
-        render json: { 'status' => 'Parameter Missing' }, status: 400
-      rescue StandardError => e
-        Appsignal.send_error(e)
-        render json: { 'status' => e.message }, status: 500
+      def set_skipped_slots_report
+        @skipped_slots_report = Report.where(
+          network: validator_params[:network],
+          name: 'build_skipped_slot_percent'
+        ).last
       end
     end
   end
