+++ config/routes.rb
@@ -117,12 +117,12 @@ Rails.application.routes.draw do
 
       # api_v1_validators GET /api/v1/validators/:network
       get 'validators/:network',
-          to: 'validators#validators_list',
+          to: 'validators#index',
           as: 'validators'
 
       # api_v1_validators GET /api/v1/validators/:network/:account
       get 'validators/:network/:account',
-          to: 'validators#validators_show',
+          to: 'validators#show',
           as: 'validator'
 
       # GET /api/v1/validator_block_history/:network/:account
